requirements.txt
gradio>=4.44.1
openai>=1.51.0
pillow>=10.4.0
requests>=2.32.3
python-dotenv>=1.0.1
